+++ cmd/root.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/fatih/structs"
+	"github.com/rebuy-de/kubernetes-deployment/pkg/api"
 	"github.com/spf13/cobra"
 )
 
@@ -18,8 +18,8 @@ func NewRootCommand() *cobra.Command {
 	debug := false
 	cmd.PersistentFlags().BoolVarP(&debug, "verbose", "v", false, "more logs")
 
-	params := new(Parameters)
-	params.Bind(cmd)
+	BindParameters(cmd)
+	params := new(api.Parameters)
 
 	cmd.PersistentPreRunE = func(cmd *cobra.Command, args []string) error {
 		log.SetLevel(log.InfoLevel)
@@ -33,7 +33,7 @@ func NewRootCommand() *cobra.Command {
 			"Commit":  BuildHash,
 		}).Infof("kubernetes-deployment started")
 
-		err := params.ReadIn()
+		err := ReadInParameters(params)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -42,9 +42,11 @@ func NewRootCommand() *cobra.Command {
 			return fmt.Errorf("You have to specify a filename.")
 		}
 
-		log.WithFields(
-			log.Fields(structs.Map(params)),
-		).Debug("config loaded")
+		log.WithFields(log.Fields{
+			"GitHubToken": fmt.Sprintf("%s****", params.GitHubToken[0:4]),
+			"Kubeconfig":  params.Kubeconfig,
+			"Filename":    params.Filename,
+		}).Debug("config loaded")
 
 		return nil
 	}
